GIX-1745: Test NnsNeuronDetail with new UI (#2951)
# Motivation

We are shuffling data in the neuron details page to make it more user
friendly.

In this PR: Add same tests for neuron detail when new UI is enabled.

In another PR: There are two pending tests that need logic in the
component and will be done in another task:
https://dfinity.atlassian.net/browse/GIX-1764

# Changes

* Add the same test cases in NnsNeuronDetail for old UI in the new UI.
* Add some methods in NnsNeuronDetailPo and NnsNeuronPageHeaderPo.

# Tests

Only test changes.

# Todos

Not worth a specific changelog entry.