GIX-1765: Manage permissions in NNS neuron buttons (#2959)
# Motivation

Some neuron actions are available only for controllers.

In this PR: Apply same rules to show/hide neuron buttons in new NNS
neuron detail.

Except one difference. The Join Neuron's fund is rendered but disabled
if neuron is not controlled by user. This is similar to what we do for
the auto stake checkbox. Currently, the whole card of Neurons' Fund
doesn't appear if neuron is not controlled by the user.

# Changes

* Hide actions if user not controller in NnsAvailableMaturityActionItem.
* Hide split button and disable join neurons' fund checkbox if user not
controller in NnsNeuronAdvancedSection.
* Hide actions if user not controller in
NnsNeuronDissolveDelayActionItem.
* Hide actions if user not controller in NnsNeuronStateItemAction.
* Add `disabled` prop to JoinCommunityFundCheckbox

# Tests

* Test that buttons are hidden when user not controller in
NnsAvailableMaturityActionItem.
* Test that split button is hidden and join neurons' fund checkbox is
disabled when user not controller in NnsNeuronAdvancedSection.
* Test that buttons are hidden when user not controller in
NnsNeuronDissolveDelayActionItem.
* Test that buttons are hidden when user not controller in
NnsNeuronStateItemAction.
* Test new `disabled` prop in JoinCommunityFundCheckbox

# Todos

This is the current functionality, applied to the new UI. It doesn't
deserver an entry in the changelog.