Proposal detail page back referrer (#2962)
# Motivation

Implementing the workaround for `onDestroy` at the top level of the
`+layout` of the detail pages (PR #2880) had for side effect to make the
navigation callbacks of its pages/layouts to not being called. As those
are not called, the dapp is not able to determine which was the
`referrer` of these pages and per extension, to effectivelly use the
referrer in back action.

Long story short, on mainnet, navigation from the "Launchpad" to a
"Proposal" and clicking back lead the users to "Proposals" page instead
of going back to "Launchpad".

# Changes

- refactor workaround in a dedicated component `LayoutNavGuard`
- use component within the layouts
- remove top layout as unused

# Notes

I noticed a type in the CHANGELOG and since I was already there, I also
improved the communication of the fix #2950 which is not related to this
PR.