Noassets feature flag (#2986)
# Motivation
There is a TODO to consolidate the way different flavours of nns-dapp
are built by generating the noassets build from a Rust feature flag.
Most of the infrastructure for flavours implemented as features was done
for the `_test` flavour so this is mostly negative code.

# Changes
* Delete the Docker stage for the noassets build.
* Define a Rust feature flag that controls whether assets are included
in the wasm.
* Add noassets to the list of build flavours.

# Tests
* Deployed the noassets build to an app subnet; it [was installed
correctly](https://github.com/dfinity/nns-dapp/actions/runs/5680641129)
and the UI, uploaded after installation, appeared as expected.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>