GIX-1777: Hotkey and NF's tags (#3003)
# Motivation

Identify hotkey neurons and those belonging to Neurons' Fund.

This is already a feature in the current neuron details page. In this
PR, it adds the same functionality in the new neuron details page.

# Changes

* New UI component HeadingTag.
* Use new HeadingTag in NnsNeuronPageHeading to render a tag when neuron
is part of Neurons' Fund or the user is a hotkey.
* Use new HeadingTag in SnsNeuronPageHeading to render a tag when the
user is a hotkey of the neuron.

# Tests

* Test new functionality in NnsNeuronPageHeading and
SnsNeuronPageHeading.

# Todos

Covered already by an entry in the changelog.