Unify Sns/Nns StakeItemAction components  (#3032)
# Motivation

The HMTL in the `NnsStakeItemAction` and `SnsStakeItemAction` components
is almost identical because these components serve the same purpose and
we want them to look the same.
Also their page objects are almost identical.

# Changes

1. Add `IncreaseStakeButton` which is used by both
`NnsIncreaseStakeButton` and `SnsIncreaseStakeButton`.
2. Add `StakeItemAction` which is used by both `NnsStakeItemAction` and
`SnsStakeItemAction`. It uses the generic `IncreaseStakeButton` and
controls the logic from the parent component.
3. Add `StakeItemActionPo` to replace both `NnsStakeItemActionPo` and
`SnsStakeItemActionPo`.

# Tests

Tests were added for `IncreaseStakeButton` and `StakeItemAction`.

# Todos

- [ ] Add entry to changelog (if necessary).

not necessary(?)