Remove empty sections in 'split-changelog' (#3052)
# Motivation

We have a script to split out the changes in "Unreleased" section in the
change log section from before and after the release.
But after using it, you still needed to manually remove the empty
section in the finished release.

# Changes

Remove empty sections automatically.

# Tests

Test is changed. You can see the effect in the diff of the commit.

# Todos

- [x] Add entry to changelog (if necessary).