GIX-1712: One confirmation with Ledger device (#3044)
# Motivation

Improve the UX of HW users.

In this PR: Users are required to confirm the transactions only ONCE
with the device. Not twice as until now.

In a nutshell, the read_state call is signed at the same time than the
other request. This is possible because the request id is deterministic
of the request and can be calculated before hand. We cache the
read_state signature so that we can reused it in the following request
call which is the read_state one.

# Changes

* LedgerIdentity uses app method "signUpdateCall" instead of "sign" that
requires the blob of the read_state call also.
* Store the read_state body and signature in the LedgerIdentity to be
reused in the following call.
* Do not use the new method if the transaction is to stake a neuron
because there is an issue in the Ledger App. This will be solved in the
next release.

# Tests

* Add a test for the LedgerIdentity to check the `transformRequest`
functionality.

# Todos

- [x] Add entry to changelog (if necessary).