E2E: Replace content instead of masking (#3072)
# Motivation

We mask the user's wallet address before taking a screenshot because it
depends on which anchor we sign in with (as well as some randomness in
the snsdemo snapshot).
We also mask the SNS names and token names because they are created
randomly in the snsdemo snapshot.

However, because we don't use a fixed-width font, the size of the masked
area still changes when the address/name changes.

Solutions considered:
1. Use a fixed size masking rectangle with a size specified in the test.
This would have to be a custom implementation because it's not supported
by Playwright.
2. Always sign in with anchor 10000. This would probably work but only
until the snsdemo snapshot is recreated.
3. Use a fixed width font for the masked areas (either for the actual
app, or only for the e2e test).
4. Mask a parent element of which the width does not depend on the
content.

Some of these might work fine but I went with:
5. Replace the contents of the relevant elements "XXXXX".

# Changes

1. Add a utility function for e2e tests to replace the content of
selected elements with something else.
2. Use it in `design.spec.ts` to replace previously masked elements'
content with "XXXXX".
3. Update screenshots.

# Tests

The screenshots look as expected.

# Todos

- [ ] Add entry to changelog (if necessary).
Covered by previous entry.