Make transactions in toy state (#2994)
# Motivation
We would like the downgrade-upgrade test to be realistic. There are two
large datasets in the state: The accounts an the transaction history.
The toy data already contains accounts but not transactions. So let's
add transactions.

# Changes
* Add a large number of toy transactions to the toy state.
* Update the downgrade-upgrade test to ensure that most transactions are
preserved.
* Note: Given that transactions are pruned then memory usage exceeds
1Gb, the number of transactions is not invariant but we should expect
most transactions to be retained.

# Tests
* The number of transactions is no longer zero:
![Screenshot from 2023-08-02
17-57-22](https://github.com/dfinity/nns-dapp/assets/5982633/171556e3-d285-45bc-bd5d-c5362d727e75)


# Todos
- Future: Adjust the number of subaccounts, hardware accounts etc per
toy account to reflect mainnet. The ratios between test and mainnet
should be close to 100%.
- [x] Add entry to changelog (if necessary).
- Already covered by "Increased the size of the persistent state in
downgrade-upgrade tests".

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>