Fix: NnsNeuronHotkeys test case (#3095)
# Motivation

Fix src/tests/lib/components/neuron-detail/NnsNeuronHotkeysCard.spec.ts

The problem was that the removeHotkey service is not called after
clicking the button. But after confirming the modal.

It was passing in this order because the mocks where not cleared and the
service was called in the previous test.

# Changes

* Use beforeEach instead of beforeAll.
* Clear mocks.
* Move check that service is called after confirmation modal.

# Tests

Only test changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Covered by an changelog entry.