Security: Bump Rust version (#3097)
# Motivation
There is a security advisory that recommends updating rust to 1.71.1.

Note: This is not super critical for us and we could just wait for the
bot to make this PR but it's a line line change so why take the risk?

Note: The bot _may_ make a PR to downgrade to 0.71.0 but we can reject
the PR.

# Changes
*  Update Rust from 1.71.0 to 1.71.1.

# Tests
CI should suffice.

# Todos

- [x] Add entry to changelog (if necessary).

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>