Fix: Randomized test sns neuron services (#3120)
# Motivation

Test sns-neuron-services.spec fails when randomizing unit tests order.

The problem was that one test was mocking a service used in another
test. Preventing from reaching the call to the api.

Solution: remove the mock and test that api is called instead.

# Changes

* In sns-neuron services remove mock to "loadNeurons". Instead, check
that "querySnsNeurons" is called.

# Tests

Only test changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Not worth an entry in the changelog.