GIX-1822: Fail gracefully on list proposal payload size error (#3128)
# Motivation

We had a problem because there were big SNS proposals and the message
limit was reached. The current UX was not the best and didn't explain
what the user could do.

In this PR: We catch that specific message and change it for a message
that the user can understand and act upon.

# Changes

* New error util "isPayloadSizeError".
* Use new util in "handleFindProposalsError" which is used in
"listProposals" service.

# Tests

* Test new error util.
* Add test cases in "listProposals" service for when api fails.

# Todos

- [ ] Add entry to changelog (if necessary).