GIX-1824: Fix LinkCanisterModal.spec (#3137)
# Motivation

Fix LinkCanisterModal.spec

The problem was that `"should attach a canister by id if name is maximum
length"` test should be failing, but it was passing because the mock was
not being cleared and the api was called in a previous test. If I
cleared to mocks in the beforeEach, the test was failing with the normal
order.

The problem with the test was that the principal was not a valid
principal. It was mixing up principalText and name. I changed so that
the name had the max length, and set a valid principalText. That fixed
the test.

I also added an expect before clicking the button that the api was not
called.

# Changes

In LinkCanisterModal.spec
* Add jest.clearAllMocks in beforeEach.
* Add a valid principal in "principalText" for `"should attach a
canister by id if name is maximum length"` and the max length name in
the `name` field.
* Add expectations of api not being called.

# Tests

Only test changes.

# Todos

- [ ] Add entry to changelog (if necessary).
Already covered by a changelog entry.