Move e2e-tests/scripts to scripts/e2e-tests (#3114)
# Motivation

All the old wdio e2e tests have been deleted.
I would like to delete the `e2e-tests` directory but there are still
scripts in it that are referenced in other places.
The scripts don't currently seem to work but fixing them is outside the
scope of migrating the e2e tests.

# Changes

1. Delete `e2e-tests/scripts/update-chromedriver` because Chromedriver
was only used by the wdio tests.
2. Move every other script in `e2e-tests/scripts/` to
`scrtips/e2e-tests`.
3. Update paths referencing the scripts.

# Tests

The scripts don't currently seem to work but fixing them is outside the
scope of migrating the e2e tests.

# Todos

- [x] Add entry to changelog (if necessary).